fix(CustomerAddress): convert Latitude and Longitude to decimal

+++ Models/Entities/CustomerAddress.cs
@@ -10,9 +10,9 @@
 
         public required byte Floor { get; set; }
 
-        public required double Latitude { get; set; }
+        public required decimal Latitude { get; set; }
 
-        public required double Longitude { get; set; }
+        public required decimal Longitude { get; set; }
 
         public string? PostalCode { get; set; }
